requirements.txt
pillow
requests
transformers
torch
tensorflow
pyautogui
keyboard
pynput